fix benchmark and chain, add parse_options in run.sh, move tacotron2_ge2e into voice_cloning
