+++ examples/aishell3/vc0/local/preprocess.sh
@@ -9,7 +9,7 @@ alignment=$3
 ge2e_ckpt_path=$4
 
 if [ ${stage} -le 0 ] && [ ${stop_stage} -ge 0 ]; then
-    python3 ${BIN_DIR}/../ge2e/inference.py \
+    python3 ${BIN_DIR}/../../ge2e/inference.py \
         --input=${input} \
         --output=${preprocess_path}/embed \
         --device="gpu" \
