+++ examples/ljspeech/voc0/run.sh
@@ -13,6 +13,11 @@ train_output_path=output
 input_mel_path=../tts0/output/test
 ckpt_name=step-10000
 
+# with the following command, you can choice the stage range you want to run
+# such as `./run.sh --stage 0 --stop-stage 0`
+# this can not be mixed use with `$1`, `$2` ...
+source ${MAIN_ROOT}/utils/parse_options.sh || exit 1
+
 if [ ${stage} -le 0 ] && [ ${stop_stage} -ge 0 ]; then
     # prepare data
     ./local/preprocess.sh ${preprocess_path} || exit -1
