+++ examples/other/ge2e/run.sh
@@ -15,6 +15,10 @@ infer_input=infer_input
 infer_output=infer_output
 ckpt_name=step-10000
 
+# with the following command, you can choice the stage range you want to run
+# such as `./run.sh --stage 0 --stop-stage 0`
+# this can not be mixed use with `$1`, `$2` ...
+source ${MAIN_ROOT}/utils/parse_options.sh || exit 1
 
 if [ ${stage} -le 0 ] && [ ${stop_stage} -ge 0 ]; then
     # prepare data
