+++ examples/vctk/voc1/run.sh
@@ -11,6 +11,11 @@ conf_path=conf/default.yaml
 train_output_path=exp/default
 ckpt_name=snapshot_iter_5000.pdz
 
+# with the following command, you can choice the stage range you want to run
+# such as `./run.sh --stage 0 --stop-stage 0`
+# this can not be mixed use with `$1`, `$2` ...
+source ${MAIN_ROOT}/utils/parse_options.sh || exit 1
+
 if [ ${stage} -le 0 ] && [ ${stop_stage} -ge 0 ]; then
     # prepare data
     ./local/preprocess.sh ${conf_path} || exit -1
