+++ tests/chains/speedyspeech/lite_train_infer.sh
