+++ tests/chains/speedyspeech/prepare.sh
