+++ tests/chains/speedyspeech/test.sh
